% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api_Registry.R
\name{APIRegistry}
\alias{APIRegistry}
\title{APIRegistry}
\description{
A registry in a module.
}
\section{Active bindings}{
\if{html}{\out{<div class="r6-active-bindings">}}
\describe{
\item{\code{module}}{(\link{APIModule})\cr
The instance the registry belongs to.}

\item{\code{name}}{(\code{character(1)})\cr
The API for the instance.}

\item{\code{class_name}}{(\code{character(1)})\cr
The class name for the registry.}

\item{\code{is_link_table}}{(\code{logical(1)})\cr
Whether the registry is a link table.}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-APIRegistry-new}{\code{APIRegistry$new()}}
\item \href{#method-APIRegistry-get}{\code{APIRegistry$get()}}
\item \href{#method-APIRegistry-df}{\code{APIRegistry$df()}}
\item \href{#method-APIRegistry-from_df}{\code{APIRegistry$from_df()}}
\item \href{#method-APIRegistry-from_path}{\code{APIRegistry$from_path()}}
\item \href{#method-APIRegistry-from_anndata}{\code{APIRegistry$from_anndata()}}
\item \href{#method-APIRegistry-get_fields}{\code{APIRegistry$get_fields()}}
\item \href{#method-APIRegistry-get_field}{\code{APIRegistry$get_field()}}
\item \href{#method-APIRegistry-get_field_names}{\code{APIRegistry$get_field_names()}}
\item \href{#method-APIRegistry-get_record_class}{\code{APIRegistry$get_record_class()}}
\item \href{#method-APIRegistry-get_temporary_record_class}{\code{APIRegistry$get_temporary_record_class()}}
\item \href{#method-APIRegistry-print}{\code{APIRegistry$print()}}
\item \href{#method-APIRegistry-to_string}{\code{APIRegistry$to_string()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-APIRegistry-new"></a>}}
\if{latex}{\out{\hypertarget{method-APIRegistry-new}{}}}
\subsection{Method \code{new()}}{
Creates an instance of this R6 class. This class should not be instantiated directly,
but rather by connecting to a LaminDB instance using the \code{\link[=api_connect]{api_connect()}} function.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{APIRegistry$new(instance, module, api, registry_name, registry_schema)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{instance}}{The instance the registry belongs to.}

\item{\code{module}}{The module the registry belongs to.}

\item{\code{api}}{The API for the instance.}

\item{\code{registry_name}}{The name of the registry.}

\item{\code{registry_schema}}{The schema for the registry.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-APIRegistry-get"></a>}}
\if{latex}{\out{\hypertarget{method-APIRegistry-get}{}}}
\subsection{Method \code{get()}}{
Get a record by ID or UID.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{APIRegistry$get(id_or_uid, include_foreign_keys = FALSE, verbose = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{id_or_uid}}{The ID or UID of the record.}

\item{\code{include_foreign_keys}}{Logical, whether to include foreign keys in the record.}

\item{\code{verbose}}{Logical, whether to print verbose output.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
A \link{APIRecord} object.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-APIRegistry-df"></a>}}
\if{latex}{\out{\hypertarget{method-APIRegistry-df}{}}}
\subsection{Method \code{df()}}{
Get a data frame summarising records in the registry
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{APIRegistry$df(limit = 100, verbose = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{limit}}{Maximum number of records to return}

\item{\code{verbose}}{Boolean, whether to print progress messages}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
A data.frame containing the available records
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-APIRegistry-from_df"></a>}}
\if{latex}{\out{\hypertarget{method-APIRegistry-from_df}{}}}
\subsection{Method \code{from_df()}}{
Create a record from a data frame
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{APIRegistry$from_df(dataframe, key = NULL, description = NULL, run = NULL)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{dataframe}}{The \code{data.frame} to create a record from}

\item{\code{key}}{A relative path within the default storage}

\item{\code{description}}{A string describing the record}

\item{\code{run}}{A \code{Run} object that creates the record}
}
\if{html}{\out{</div>}}
}
\subsection{Details}{
Creating records is only possible for the default instance, requires the
Python \code{lamindb} module and is only implemented for the core \code{Artifact}
registry.
}

\subsection{Returns}{
A \code{TemporaryRecord} object containing the new record. This is not
saved to the database until \code{temp_record$save()} is called.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-APIRegistry-from_path"></a>}}
\if{latex}{\out{\hypertarget{method-APIRegistry-from_path}{}}}
\subsection{Method \code{from_path()}}{
Create a record from a path
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{APIRegistry$from_path(path, key = NULL, description = NULL, run = NULL)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{path}}{Path to create a record from}

\item{\code{key}}{A relative path within the default storage}

\item{\code{description}}{A string describing the record}

\item{\code{run}}{A \code{Run} object that creates the record}
}
\if{html}{\out{</div>}}
}
\subsection{Details}{
Creating records is only possible for the default instance, requires the
Python \code{lamindb} module and is only implemented for the core \code{Artifact}
registry.
}

\subsection{Returns}{
A \code{TemporaryRecord} object containing the new record. This is not
saved to the database until \code{temp_record$save()} is called.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-APIRegistry-from_anndata"></a>}}
\if{latex}{\out{\hypertarget{method-APIRegistry-from_anndata}{}}}
\subsection{Method \code{from_anndata()}}{
Create a record from an \code{AnnData}
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{APIRegistry$from_anndata(adata, key = NULL, description = NULL, run = NULL)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{adata}}{The \link[anndata:AnnData]{anndata::AnnData} object to create a record from}

\item{\code{key}}{A relative path within the default storage}

\item{\code{description}}{A string describing the record}

\item{\code{run}}{A \code{Run} object that creates the record}
}
\if{html}{\out{</div>}}
}
\subsection{Details}{
Creating records is only possible for the default instance, requires the
Python \code{lamindb} module and is only implemented for the core \code{Artifact}
registry.
}

\subsection{Returns}{
A \code{TemporaryRecord} object containing the new record. This is not
saved to the database until \code{temp_record$save()} is called.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-APIRegistry-get_fields"></a>}}
\if{latex}{\out{\hypertarget{method-APIRegistry-get_fields}{}}}
\subsection{Method \code{get_fields()}}{
Get the fields in the registry.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{APIRegistry$get_fields()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
A list of \link{APIField} objects.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-APIRegistry-get_field"></a>}}
\if{latex}{\out{\hypertarget{method-APIRegistry-get_field}{}}}
\subsection{Method \code{get_field()}}{
Get a field by name.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{APIRegistry$get_field(field_name)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{field_name}}{The name of the field.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
A \link{APIField} object.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-APIRegistry-get_field_names"></a>}}
\if{latex}{\out{\hypertarget{method-APIRegistry-get_field_names}{}}}
\subsection{Method \code{get_field_names()}}{
Get the field names in the registry.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{APIRegistry$get_field_names()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
A character vector of field names.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-APIRegistry-get_record_class"></a>}}
\if{latex}{\out{\hypertarget{method-APIRegistry-get_record_class}{}}}
\subsection{Method \code{get_record_class()}}{
Get the record class for the registry.

Note: This method is intended for internal use only and may be removed in the future.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{APIRegistry$get_record_class()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
A \link{APIRecord} class.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-APIRegistry-get_temporary_record_class"></a>}}
\if{latex}{\out{\hypertarget{method-APIRegistry-get_temporary_record_class}{}}}
\subsection{Method \code{get_temporary_record_class()}}{
Get the temporary record class for the registry.

Note: This method is intended for internal use only and may be removed in the future.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{APIRegistry$get_temporary_record_class()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
A \code{TemporaryRecord} class.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-APIRegistry-print"></a>}}
\if{latex}{\out{\hypertarget{method-APIRegistry-print}{}}}
\subsection{Method \code{print()}}{
Print a \code{APIRegistry}
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{APIRegistry$print(style = TRUE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{style}}{Logical, whether the output is styled using ANSI codes}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
A character vector
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-APIRegistry-to_string"></a>}}
\if{latex}{\out{\hypertarget{method-APIRegistry-to_string}{}}}
\subsection{Method \code{to_string()}}{
Create a string representation of a \code{APIRegistry}
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{APIRegistry$to_string(style = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{style}}{Logical, whether the output is styled using ANSI codes}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
A \code{cli::cli_ansi_string} if \code{style = TRUE} or a character vector
}
}
}
