% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api_RelatedRecords.R
\name{APIRelatedRecords}
\alias{APIRelatedRecords}
\title{APIRelatedRecords}
\description{
A container for accessing records with a one-to-many or many-to-many
relationship.
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-APIRelatedRecords-new}{\code{APIRelatedRecords$new()}}
\item \href{#method-APIRelatedRecords-df}{\code{APIRelatedRecords$df()}}
\item \href{#method-APIRelatedRecords-print}{\code{APIRelatedRecords$print()}}
\item \href{#method-APIRelatedRecords-to_string}{\code{APIRelatedRecords$to_string()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-APIRelatedRecords-new"></a>}}
\if{latex}{\out{\hypertarget{method-APIRelatedRecords-new}{}}}
\subsection{Method \code{new()}}{
Creates an instance of this R6 class. This class should not be instantiated directly,
but rather by connecting to a LaminDB instance using the \code{\link[=api_connect]{api_connect()}} function.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{APIRelatedRecords$new(instance, registry, field, related_to, api)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{instance}}{The instance the records list belongs to.}

\item{\code{registry}}{The registry the records list belongs to.}

\item{\code{field}}{The field associated with the records list.}

\item{\code{related_to}}{ID or UID of the parent that records are related to.}

\item{\code{api}}{The API for the instance.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-APIRelatedRecords-df"></a>}}
\if{latex}{\out{\hypertarget{method-APIRelatedRecords-df}{}}}
\subsection{Method \code{df()}}{
Get a data frame summarising records in the registry
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{APIRelatedRecords$df(limit = 100, verbose = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{limit}}{Maximum number of records to return}

\item{\code{verbose}}{Boolean, whether to print progress messages}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
A data.frame containing the available records
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-APIRelatedRecords-print"></a>}}
\if{latex}{\out{\hypertarget{method-APIRelatedRecords-print}{}}}
\subsection{Method \code{print()}}{
Print a \code{APIRelatedRecords}
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{APIRelatedRecords$print(style = TRUE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{style}}{Logical, whether the output is styled using ANSI codes}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-APIRelatedRecords-to_string"></a>}}
\if{latex}{\out{\hypertarget{method-APIRelatedRecords-to_string}{}}}
\subsection{Method \code{to_string()}}{
Create a string representation of a \code{APIRelatedRecords}
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{APIRelatedRecords$to_string(style = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{style}}{Logical, whether the output is styled using ANSI codes}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
A \code{cli::cli_ansi_string} if \code{style = TRUE} or a character vector
}
}
}
