% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/connect.R
\name{connect}
\alias{connect}
\title{Connect to instance}
\usage{
connect(slug)
}
\arguments{
\item{slug}{The instance slug \code{account_handle/instance_name} or URL.
If the instance is owned by you, it suffices to pass the instance name.}
}
\description{
Connect to instance
}
\examples{
\dontrun{
  # first run 'lamin login' to authenticate
  instance <- connect("laminlabs/cellxgene")
  instance
}
}
