% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{get_current_lamin_instance}
\alias{get_current_lamin_instance}
\title{Get current LaminDB instance}
\usage{
get_current_lamin_instance(ignore_none = TRUE, silent = FALSE)
}
\arguments{
\item{ignore_none}{Whether to ignore the \code{"none/none"} virtual instance as a
valid instance and return \code{NULL}}

\item{silent}{Whether to suppress messages}
}
\value{
The slug of the current LaminDB instance, or \code{NULL} invisibly if no
instance is found
}
\description{
Get the currently connected LaminDB instance
}
