% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{get_current_lamin_settings}
\alias{get_current_lamin_settings}
\title{Get current LaminDB settings}
\usage{
get_current_lamin_settings(minimal = FALSE, silent = FALSE)
}
\arguments{
\item{minimal}{If \code{TRUE}, quickly extract a minimal list of important
settings instead of converting the complete settings object}

\item{silent}{Whether to suppress messages}
}
\value{
A list of the current LaminDB settings
}
\description{
Get the current LaminDB settings as an R list
}
