% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{get_current_lamin_user}
\alias{get_current_lamin_user}
\title{Get current LaminDB user}
\usage{
get_current_lamin_user()
}
\value{
The handle of the current LaminDB user, or \code{NULL} invisibly if no
user is found
}
\description{
Get the currently logged in LaminDB user
}
\details{
This is done via \code{\link[=lamin_settings]{lamin_settings()}} to avoid importing Python \code{lamindb}
}
