% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{get_lamin_settings}
\alias{get_lamin_settings}
\title{Get Lamin settings}
\usage{
get_lamin_settings()
}
\value{
A list of the current LaminDB settings
}
\description{
Get the current LaminDB settings as an R list
}
\details{
This is done using \code{\link[callr:r]{callr::r()}} to avoid importing Python \code{lamindb} in the
global environment
}
