% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import.R
\name{import_bionty}
\alias{import_bionty}
\title{Import bionty}
\usage{
import_bionty()
}
\value{
An object representing the \code{bionty} Python package
}
\description{
Import the \code{bionty} Python package
}
\examples{
\dontrun{
bt <- import_bionty()
}
}
