% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import.R
\name{import_lamindb}
\alias{import_lamindb}
\title{Import lamindb}
\usage{
import_lamindb()
}
\value{
An object representing the \code{lamindb} Python package
}
\description{
Import the \code{lamindb} Python package
}
\examples{
\dontrun{
ln <- import_lamindb()
}
}
