% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lamindb.R, R/import.R
\name{import_lamindb}
\alias{import_lamindb}
\alias{import_bionty}
\alias{import_wetlab}
\alias{import_clinicore}
\alias{importing}
\alias{import_module}
\title{Import Python modules}
\usage{
import_lamindb()

import_bionty()

import_wetlab()

import_clinicore()

import_module(module, check_instance = FALSE)
}
\arguments{
\item{module}{The name of the Python module to import}

\item{check_instance}{Whether to check if the package is installed in the
current LaminDB instance.}
}
\value{
An object representing a Python package
}
\description{
These function can be used to import LaminDB Python modules with additional
checks and nicer error messages.
}
\details{
\itemize{
\item \code{import_lamindb()} - Import the main \code{lamindb} package. This is the
starting point for interacting with a LaminDB instance.
\item \code{import_bionty()} - Import the \code{bionty} module for accessing biological
entities.
\item \code{import_wetlab()} - Import the \code{wetlab} module for accessing wet lab
entities.
\item \code{import_clinicore()} - Import the \code{clincore} module for accessing clinical
entities.
\item \code{import_module()} - Import any Python module. This is mostly intended for
internal use and should only be used if there is not a specific function
for that module.
}
}
\examples{
\dontrun{
# Import lamindb to start interacting with an instance
ln <- import_lamindb()

# Import other LaminDB modules
bt <- import_bionty()
wl <- import_wetlab()
cc <- import_clinicore()

# Import any Python module
np <- import_module("numpy")
}
}
