% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lamin_cli.R
\name{lamin_connect}
\alias{lamin_connect}
\title{Connect to a LaminDB instance}
\usage{
lamin_connect(instance)
}
\arguments{
\item{instance}{Either a slug giving the instance to connect to
(\verb{<owner>/<name>}) or an instance URL (\verb{https://lamin.ai/owner/name})}
}
\description{
Connect to a LaminDB instance by calling \verb{lamin connect} on the command line
}
\details{
Running this will set the LaminDB auto-connect option to \code{True} so you
auto-connect to \code{instance} when importing Python \code{lamindb}.
}
\examples{
\dontrun{
lamin_connect("laminlabs/cellxgene")
}
}
