% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lamin_cli.R
\name{lamin_delete}
\alias{lamin_delete}
\title{LaminDB delete}
\usage{
lamin_delete(instance, force = FALSE)
}
\arguments{
\item{instance}{Identifier for the instance to delete (e.g. "owner/name")}

\item{force}{Whether to force deletion without asking for confirmation}
}
\description{
Delete a LaminDB entity. Currently only supports instances.
}
\examples{
\dontrun{
lamin_init("to-delete")
lamin_delete("to-delete")
}
}
