% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lamin_cli.R
\name{lamin_disconnect}
\alias{lamin_disconnect}
\title{Disconnect from a LaminDB instance}
\usage{
lamin_disconnect()
}
\description{
Disconnect from the current LaminDB instance by calling \verb{lamin connect} on
the command line
}
\examples{
\dontrun{
lamin_disconnect()
}
}
