% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/connect.R
\name{lamin_login}
\alias{lamin_login}
\title{Login to LaminDB}
\usage{
lamin_login(user = NULL, api_key = NULL)
}
\arguments{
\item{user}{Handle for the user to login as}

\item{api_key}{API key for a user}
}
\description{
Login as a LaminDB user
}
\details{
Setting \code{user} will run \verb{lamin login <user>}. Setting \code{api_key} will set the
\code{LAMIN_API_KEY} environment variable temporarily with \code{withr::with_envvar()}
and run \verb{lamin login}. If neither \code{user} or \code{api_key} are set the user handle
will be retrieved from the user settings file. If that is not possible
\verb{lamin login} will only be run if \code{LAMIN_API_KEY} is set.
}
