% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lamin_cli.R
\name{lamin_save}
\alias{lamin_save}
\title{Save to a LaminDB instance}
\usage{
lamin_save(filepath, key = NULL, description = NULL, registry = NULL)
}
\arguments{
\item{filepath}{Path to the file or folder to save}

\item{key}{The key for the saved item}

\item{description}{The description for the saved item}

\item{registry}{The registry for the saved item}
}
\description{
Save a file or folder to a LaminDB instance by calling \verb{lamin save} on the
command line
}
\details{
See \verb{lamin save --help} for details of what database entries are created for
different file types
}
\examples{
\dontrun{
my_file <- tempfile()
lamin_save(my_file)
}
}
