% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lamin_cli.R
\name{lamin_settings}
\alias{lamin_settings}
\title{Lamin settings}
\usage{
lamin_settings()
}
\value{
A list of the current LaminDB settings
}
\description{
Get the current LaminDB settings by calling \verb{lamin settings} on the command
line
}
\examples{
\dontrun{
lamin_settings()
}
}
