% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LaminDB.R
\name{LaminDB}
\alias{LaminDB}
\title{LaminDB class}
\description{
This class is used to interact with the LaminDB API.
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{types}}{Object types defined by the LaminDB schema
Initialize the LaminDB class}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-LaminDB-new}{\code{LaminDB$new()}}
\item \href{#method-LaminDB-clone}{\code{LaminDB$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-LaminDB-new"></a>}}
\if{latex}{\out{\hypertarget{method-LaminDB-new}{}}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{LaminDB$new(url, instance_id, schema_id)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{url}}{The base URL of the LaminDB API.}

\item{\code{instance_id}}{The instance ID of the LaminDB instance.}

\item{\code{schema_id}}{The schema ID of the LaminDB schema.
importFrom dplyr filter \%>\%
importFrom purrr pmap}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-LaminDB-clone"></a>}}
\if{latex}{\out{\hypertarget{method-LaminDB-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{LaminDB$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
