% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sitrep.R
\name{laminr_sitrep}
\alias{laminr_sitrep}
\title{laminr situation report}
\usage{
laminr_sitrep()
}
\value{
Prints details of the current laminr status
}
\description{
Overview of the current status of the laminr package and its dependencies.
Can be useful for debugging.
}
\details{
Provides information that can be useful for debugging. To run the function
when an error occurs set \code{options(error = laminr::laminr_sitrep)}. Note that
this should be used with some caution as it will print the status whenever
any error occurs.
}
\examples{
laminr_sitrep()
}
