% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RecordsList.R
\name{RecordsList}
\alias{RecordsList}
\title{RecordsList}
\description{
A container for accessing several records of one type.
}
\section{Active bindings}{
\if{html}{\out{<div class="r6-active-bindings">}}
\describe{
\item{\code{field}}{(\link{Field})\cr
The field the records are related to.}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-RecordsList-new}{\code{RecordsList$new()}}
\item \href{#method-RecordsList-df}{\code{RecordsList$df()}}
\item \href{#method-RecordsList-print}{\code{RecordsList$print()}}
\item \href{#method-RecordsList-to_string}{\code{RecordsList$to_string()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-RecordsList-new"></a>}}
\if{latex}{\out{\hypertarget{method-RecordsList-new}{}}}
\subsection{Method \code{new()}}{
Creates an instance of this R6 class. This class should not be instantiated directly,
but rather by connecting to a LaminDB instance using the \code{\link[=connect]{connect()}} function.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{RecordsList$new(instance, registry, field, parent, api)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{instance}}{The instance the records list belongs to.}

\item{\code{registry}}{The registry the records list belongs to.}

\item{\code{field}}{The field associated with the records list.}

\item{\code{parent}}{ID or UID of the parent record the records list belongs to.}

\item{\code{api}}{The API for the instance.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-RecordsList-df"></a>}}
\if{latex}{\out{\hypertarget{method-RecordsList-df}{}}}
\subsection{Method \code{df()}}{
Get a data frame summarising records in the registry
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{RecordsList$df(limit = 100, verbose = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{limit}}{Maximum number of records to return}

\item{\code{verbose}}{Boolean, whether to print progress messages}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
A data.frame containing the available records
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-RecordsList-print"></a>}}
\if{latex}{\out{\hypertarget{method-RecordsList-print}{}}}
\subsection{Method \code{print()}}{
Print a \code{RecordsList}
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{RecordsList$print(style = TRUE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{style}}{Logical, whether the output is styled using ANSI codes}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-RecordsList-to_string"></a>}}
\if{latex}{\out{\hypertarget{method-RecordsList-to_string}{}}}
\subsection{Method \code{to_string()}}{
Create a string representation of a \code{RecordsList}
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{RecordsList$to_string(style = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{style}}{Logical, whether the output is styled using ANSI codes}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
A \code{cli::cli_ansi_string} if \code{style = TRUE} or a character vector
}
}
}
