% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/connect.R
\name{connect}
\alias{connect}
\title{Connect to a Lamin instance}
\usage{
connect(slug = NULL)
}
\arguments{
\item{slug}{The slug of the instance to connect to. If not provided, the current instance is used.}
}
\description{
Connect to a Lamin instance using the current instance settings.
}
\examples{
\dontrun{
options(
  lamindb_current_instance = list(
    owner = "lamin",
    name = "example",
    api_url = "https://us-west-2.api.lamin.ai",
    id = "0123456789abcdefghijklmnopqrstuv",
    schema_id = "0123456789abcdefghijklmnopqrstuv"
  )
)
db <- connect()
}
}
