% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{get_current_lamin_instance}
\alias{get_current_lamin_instance}
\title{Get current LaminDB instance}
\usage{
get_current_lamin_instance()
}
\value{
The slug of the current LaminDB instance, or \code{NULL} invisibly if no
instance is found
}
\description{
Get the currently connected LaminDB instance
}
\details{
This is done via a \code{\link[=lamin_settings]{lamin_settings()}} to avoid importing Python \code{lamindb}
}
