% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import.R
\name{import_module}
\alias{import_module}
\title{Import Python modules}
\usage{
import_module(module)
}
\arguments{
\item{module}{The name of the Python module to import}
}
\value{
An object representing a Python package
}
\description{
This function can be used to import LaminDB Python modules with additional
checks and nicer error messages.
}
\examples{
\dontrun{
# Import lamindb to start interacting with an instance
ln <- import_module("lamindb")

# Import other LaminDB modules
bt <- import_module("bionty")
wl <- import_module("wetlab")
cc <- import_module("clincore")

# Import any Python module
np <- import_module("numpy")
}
}
