% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Instance.R
\name{Instance}
\alias{Instance}
\title{Instance}
\description{
Connect to a LaminDB instance using the \code{\link[=connect]{connect()}} function.
The instance object provides access to the modules and registries
of the instance.
}
\details{
Note that by connecting to an instance via \code{\link[=connect]{connect()}}, you receive
a "richer" version of the Instance class documented here, providing
direct access to all core registries and additional modules.
See the vignette on "Package Architecture" for more information:
\code{vignette("architecture", package = "laminr")}.
}
\examples{
\dontrun{
# Connect to an instance
db <- connect("laminlabs/cellxgene")

# fetch an artifact
artifact <- db$Artifact$get("KBW89Mf7IGcekja2hADu")

# describe the artifact
artifact$describe()

# view field
artifact$id

# load dataset
artifact$load()
}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-Instance-new}{\code{Instance$new()}}
\item \href{#method-Instance-get_modules}{\code{Instance$get_modules()}}
\item \href{#method-Instance-get_module}{\code{Instance$get_module()}}
\item \href{#method-Instance-get_module_names}{\code{Instance$get_module_names()}}
\item \href{#method-Instance-get_settings}{\code{Instance$get_settings()}}
\item \href{#method-Instance-get_api}{\code{Instance$get_api()}}
\item \href{#method-Instance-print}{\code{Instance$print()}}
\item \href{#method-Instance-to_string}{\code{Instance$to_string()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Instance-new"></a>}}
\if{latex}{\out{\hypertarget{method-Instance-new}{}}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Instance$new(settings, api, schema)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{settings}}{The settings for the instance}

\item{\code{api}}{The API for the instance}

\item{\code{schema}}{The schema for the instance
Get the modules for the instance.
Get a module by name.
Get the names of the modules. Example: \code{c("core", "bionty")}.
Get instance settings.
Get instance API.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Instance-get_modules"></a>}}
\if{latex}{\out{\hypertarget{method-Instance-get_modules}{}}}
\subsection{Method \code{get_modules()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Instance$get_modules()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Instance-get_module"></a>}}
\if{latex}{\out{\hypertarget{method-Instance-get_module}{}}}
\subsection{Method \code{get_module()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Instance$get_module(module_name)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Instance-get_module_names"></a>}}
\if{latex}{\out{\hypertarget{method-Instance-get_module_names}{}}}
\subsection{Method \code{get_module_names()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Instance$get_module_names()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Instance-get_settings"></a>}}
\if{latex}{\out{\hypertarget{method-Instance-get_settings}{}}}
\subsection{Method \code{get_settings()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Instance$get_settings()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Instance-get_api"></a>}}
\if{latex}{\out{\hypertarget{method-Instance-get_api}{}}}
\subsection{Method \code{get_api()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Instance$get_api()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Instance-print"></a>}}
\if{latex}{\out{\hypertarget{method-Instance-print}{}}}
\subsection{Method \code{print()}}{
Print an \code{Instance}
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Instance$print(style = TRUE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{style}}{Logical, whether the output is styled using ANSI codes}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Instance-to_string"></a>}}
\if{latex}{\out{\hypertarget{method-Instance-to_string}{}}}
\subsection{Method \code{to_string()}}{
Create a string representation of an \code{Instance}
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Instance$to_string(style = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{style}}{Logical, whether the output is styled using ANSI codes}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
A \code{cli::cli_ansi_string} if \code{style = TRUE} or a character vector
}
}
}
