% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lamin_cli.R
\name{lamin_init}
\alias{lamin_init}
\title{Initalise LaminDB}
\usage{
lamin_init(storage, name = NULL, db = NULL, modules = NULL)
}
\arguments{
\item{storage}{A local directory, AWS S3 bucket or Google Cloud Storage bucket}

\item{name}{A name for the instance}

\item{db}{A Postgres database connection URL, use \code{NULL} for SQLite.}

\item{modules}{A vector of modules to include (e.g. "bionty")}
}
\description{
Initialise a new LaminDB instance
}
\examples{
\dontrun{
lamin_init("mydata", modules = c("bionty", "wetlab"))
}
}
