% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/laminr-package.R
\docType{package}
\name{laminr-package}
\alias{laminr}
\alias{laminr-package}
\title{laminr: LaminDB interface in R}
\description{
The laminr package provides an interface to the LaminDB database. It allows you to query the database and download data from it.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/laminlabs/laminr}
  \item Report bugs at \url{https://github.com/laminlabs/laminr/issues}
}

}
\author{
\strong{Maintainer}: Robrecht Cannoodt \email{rcannood@gmail.com} (\href{https://orcid.org/0000-0003-3641-729X}{ORCID})

Authors:
\itemize{
  \item Lamin Labs \email{open-source@lamin.ai}
}

}
\keyword{internal}
