% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/laminr-package.R
\docType{package}
\name{laminr-package}
\alias{laminr}
\alias{laminr-package}
\title{laminr: 'LaminDB' Interface in R}
\description{
Interact with 'LaminDB' from R. 'LaminDB' is an open-source data framework for biology. This package allows you to query and download data from 'LaminDB' instances.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://laminr.lamin.ai}
  \item \url{https://github.com/laminlabs/laminr}
  \item Report bugs at \url{https://github.com/laminlabs/laminr/issues}
}

}
\author{
\strong{Maintainer}: Robrecht Cannoodt \email{robrecht@data-intuitive.com} (\href{https://orcid.org/0000-0003-3641-729X}{ORCID}) (rcannood)

Authors:
\itemize{
  \item Luke Zappia \email{luke@data-intuitive.com} (\href{https://orcid.org/0000-0001-7744-8565}{ORCID}) (lazappi)
  \item Lamin Labs \email{open-source@lamin.ai} [copyright holder]
}

}
\keyword{internal}
