% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Registry.R
\name{Registry}
\alias{Registry}
\title{Registry}
\description{
A registry in a module.
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-Registry-new}{\code{Registry$new()}}
\item \href{#method-Registry-get}{\code{Registry$get()}}
\item \href{#method-Registry-get_fields}{\code{Registry$get_fields()}}
\item \href{#method-Registry-get_field}{\code{Registry$get_field()}}
\item \href{#method-Registry-get_field_names}{\code{Registry$get_field_names()}}
\item \href{#method-Registry-get_record_class}{\code{Registry$get_record_class()}}
\item \href{#method-Registry-print}{\code{Registry$print()}}
\item \href{#method-Registry-to_string}{\code{Registry$to_string()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Registry-new"></a>}}
\if{latex}{\out{\hypertarget{method-Registry-new}{}}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Registry$new(instance, module, api, registry_name, registry_schema)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{instance}}{The instance the registry belongs to.}

\item{\code{module}}{The module the registry belongs to.}

\item{\code{api}}{The API for the instance.}

\item{\code{registry_name}}{The name of the registry.}

\item{\code{registry_schema}}{The schema for the registry.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Registry-get"></a>}}
\if{latex}{\out{\hypertarget{method-Registry-get}{}}}
\subsection{Method \code{get()}}{
Get a record by ID or UID.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Registry$get(id_or_uid, include_foreign_keys = FALSE, verbose = FALSE)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Registry-get_fields"></a>}}
\if{latex}{\out{\hypertarget{method-Registry-get_fields}{}}}
\subsection{Method \code{get_fields()}}{
Get the fields in the registry.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Registry$get_fields()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Registry-get_field"></a>}}
\if{latex}{\out{\hypertarget{method-Registry-get_field}{}}}
\subsection{Method \code{get_field()}}{
Get a field by name.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Registry$get_field(field_name)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Registry-get_field_names"></a>}}
\if{latex}{\out{\hypertarget{method-Registry-get_field_names}{}}}
\subsection{Method \code{get_field_names()}}{
Get the field names in the registry.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Registry$get_field_names()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Registry-get_record_class"></a>}}
\if{latex}{\out{\hypertarget{method-Registry-get_record_class}{}}}
\subsection{Method \code{get_record_class()}}{
Get the record class for the registry.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Registry$get_record_class()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Registry-print"></a>}}
\if{latex}{\out{\hypertarget{method-Registry-print}{}}}
\subsection{Method \code{print()}}{
Print a \code{Registry}
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Registry$print(style = TRUE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{style}}{Logical, whether the output is styled using ANSI codes}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Registry-to_string"></a>}}
\if{latex}{\out{\hypertarget{method-Registry-to_string}{}}}
\subsection{Method \code{to_string()}}{
Create a string representation of a \code{Registry}
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Registry$to_string(style = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{style}}{Logical, whether the output is styled using ANSI codes}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
A \code{cli::cli_ansi_string} if \code{style = TRUE} or a character vector
}
}
}
